+++ test/widgets/content_test.dart
@@ -25,7 +25,6 @@ import '../model/binding.dart';
 import '../model/content_test.dart';
 import '../model/store_checks.dart';
 import '../model/test_store.dart';
-import '../stdlib_checks.dart';
 import '../test_images.dart';
 import '../test_navigation.dart';
 import 'checks.dart';
@@ -1281,69 +1280,6 @@ void main() {
       await tester.pumpWidget(
         RealmContentNetworkImage(Uri.parse('https://zulip.invalid/path/to/image.png'), filterQuality: FilterQuality.medium));
       check(tester.takeException()).isA<AssertionError>();
-    });
-  });
-
-  group('AvatarImage', () {
-    late PerAccountStore store;
-
-    Future<Uri?> actualUrl(WidgetTester tester, String avatarUrl, [double? size]) async {
-      addTearDown(testBinding.reset);
-      await testBinding.globalStore.add(eg.selfAccount, eg.initialSnapshot());
-      store = await testBinding.globalStore.perAccount(eg.selfAccount.id);
-      final user = eg.user(avatarUrl: avatarUrl);
-      await store.addUser(user);
-
-      prepareBoringImageHttpClient();
-      await tester.pumpWidget(GlobalStoreWidget(
-        child: PerAccountStoreWidget(accountId: eg.selfAccount.id,
-          child: AvatarImage(userId: user.userId, size: size ?? 30))));
-      await tester.pump();
-      await tester.pump();
-      tester.widget(find.byType(AvatarImage));
-      final widgets = tester.widgetList<RealmContentNetworkImage>(
-        find.byType(RealmContentNetworkImage));
-      return widgets.firstOrNull?.src;
-    }
-
-    testWidgets('smoke with absolute URL', (tester) async {
-      const avatarUrl = 'https://example/avatar.png';
-      check(await actualUrl(tester, avatarUrl)).isNotNull()
-        .asString.equals(avatarUrl);
-      debugNetworkImageHttpClientProvider = null;
-    });
-
-    testWidgets('smoke with relative URL', (tester) async {
-      const avatarUrl = '/avatar.png';
-      check(await actualUrl(tester, avatarUrl))
-        .equals(store.tryResolveUrl(avatarUrl)!);
-      debugNetworkImageHttpClientProvider = null;
-    });
-
-   testWidgets('absolute URL, larger size', (tester) async {
-      tester.view.devicePixelRatio = 2.5;
-      addTearDown(tester.view.resetDevicePixelRatio);
-
-      const avatarUrl = 'https://example/avatar.png';
-      check(await actualUrl(tester, avatarUrl, 50)).isNotNull()
-        .asString.equals(avatarUrl.replaceAll('.png', '-medium.png'));
-      debugNetworkImageHttpClientProvider = null;
-    });
-
-    testWidgets('relative URL, larger size', (tester) async {
-      tester.view.devicePixelRatio = 2.5;
-      addTearDown(tester.view.resetDevicePixelRatio);
-
-      const avatarUrl = '/avatar.png';
-      check(await actualUrl(tester, avatarUrl, 50))
-        .equals(store.tryResolveUrl('/avatar-medium.png')!);
-      debugNetworkImageHttpClientProvider = null;
-    });
-
-    testWidgets('smoke with invalid URL', (tester) async {
-      const avatarUrl = '::not a URL::';
-      check(await actualUrl(tester, avatarUrl)).isNull();
-      debugNetworkImageHttpClientProvider = null;
     });
   });
